refactor: extract active message system to standalone dynamo-am crate
Migrated the active message system from lib/runtime to a new standalone
crate at lib/am (dynamo-am) to improve modularity and maintainability.

- Created new dynamo-am workspace crate with all active message code
- Updated runtime to re-export dynamo-am as active_message
- Moved all tests, examples, and documentation to new crate
- Fixed import paths throughout to use crate root instead of nested paths
- Enabled serde feature for bytes dependency
- All 5 integration tests passing in new location
