+++ Cargo.toml
@@ -5,6 +5,7 @@
 members = [
     "components/metrics",
     "launch/*",
+    "lib/am",
     "lib/llm",
     "lib/runtime",
     "lib/tokens",
@@ -27,6 +28,7 @@ keywords = ["llm", "genai", "inference", "nvidia", "distributed", "dynamo"]
 
 [workspace.dependencies]
 # Local crates
+dynamo-am = { path = "lib/am", version = "0.5.0" }
 dynamo-runtime = { path = "lib/runtime", version = "0.5.0" }
 dynamo-llm = { path = "lib/llm", version = "0.5.0" }
 dynamo-tokens = { path = "lib/tokens", version = "0.5.0" }
