+++ lib/am/src/network_client.rs
