+++ lib/runtime/src/lib.rs
@@ -33,7 +33,9 @@ use async_once_cell::OnceCell;
 mod config;
 pub use config::RuntimeConfig;
 
-pub mod active_message;
+// Re-export active message from dynamo-am crate
+pub use dynamo_am as active_message;
+
 pub mod component;
 pub mod discovery;
 pub mod engine;
